requirements.txt
transformers
torch
faiss-cpu
wikipedia
gradio
python-docx
wikipedia-api
PyPDF2
nltk